dbmf/views.py
from django.urls import reverse_lazy
from django.views import generic

from bootstrap_modal_forms.generic import (BSModalCreateView,
                                           BSModalUpdateView,
                                           BSModalReadView,
                                           BSModalDeleteView)


class Dbmf(generic.ListView):
    template_name = 'dbmf/dbmf.html'


