requirements.txt
Django==2.1.7
django-bootstrap-modal-forms==1.4.2
django-environ==0.4.5
django-extra-views==0.12.0
django-funky-sheets==0.1.2
django-widget-tweaks==1.4.3
gunicorn==19.9.0
pytz==2018.9
six==1.12.0
-i https://pypi.python.org/simple
django-authtools==1.6.0
django-braces==1.13.0
django-crispy-forms==1.7.2
django-environ==0.4.5
easy-thumbnails==2.5
pillow==5.3.0
django-debug-toolbar==1.10.1
sqlparse==0.2.4
werkzeug==0.14.1
django[argon2]